Disable "git prompt"
I left the git ENV variables intact so if I get the urge to re-enable
the git status prompt, I can. But it's become a nuisance, and I still
find myself typing 'git st', which is an alias for git status -sb, and
gives the same information for a clean tree and not an overwhelmingly
greater amount for a dirty tree. I actually prefer it a lot.

And of course, I was sick of a third of every line being taken up by the
prompt, and I'm not going to use a two-line prompt. That's just wrong.