Force statusline to redraw on VimEnter
Sometimes when vim first starts up, the statusline is missing and the
cursor is not positioned in the correct place. This only happens when
'lazyredraw' is set, and according to the internet, only in iTerm2 or
Apple Terminal.

When this happens, it can be corrected by interacting with the window,
or triggering FocusGained, if vim is configured to receive focus events
from the terminal.

Turning off 'lazyredraw' is a non-starter, so the best solution seems to
be invoking redrawstatus  on VimEnter.